Implementation of debugDescription: Minor cleanup
