+++ Tests/OktaIdxTests/IDXClientTests.swift
@@ -36,7 +36,6 @@ class IDXClientTests: XCTestCase {
 
     func testConstructors() {
         let idx = IDXClient(context: context, api: api)
-        print(context.debugDescription)
         
         XCTAssertNotNil(idx)
         XCTAssertEqual(idx.context, context)
